Converted to Express Http Server - Pass main intereface through server to fix rendering issues - Move AppData and GameData to HttpData to perserve file structure
