+++ .gitignore
@@ -1,0 +1,21 @@
+# Node modules
+node_modules/
+
+# Electron build output
+TardQuest-win32-x64/
+
+# OS-specific files
+.DS_Store
+Thumbs.db
+
+# Logs
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+
+# Environment files
+.env
+.env.local
+.env.*.local
+
+# VSCode settings
