--- AppData/retro-filter.css
@@ -1,70 +1,0 @@
-/* Create this file: retro-filter.css */
-.game-container {
-  position: relative;
-  /* Remove fixed width and height - let it be controlled by the HTML/JS */
-}
-
-.game-container iframe {
-  width: 100%;
-  height: 100%;
-  border: none;
-  display: block;
-  position: relative;
-  z-index: 1;
-}
-
-/* Scanlines and dithering disabled by default */
-.game-container::before,
-.game-container::after {
-  content: none;
-}
-
-/* Scanlines */
-.game-container.crt-enabled::before {
-  content: "";
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  pointer-events: none;
-  z-index: 2;
-  background: repeating-linear-gradient(
-    to bottom,
-    rgba(var(--scanline-color), 0.07),
-    rgba(var(--scanline-color), 0.07) 1px,
-    transparent 3px,
-    transparent 4px
-  );
-  animation: flicker 0.15s infinite;
-}
-
-/* Glow and blur overlay */
-.game-container.crt-enabled .crt-glow {
-  pointer-events: none;
-  position: absolute;
-  top: 0; 
-  left: 0;
-  width: 100%; 
-  height: 100%;
-  z-index: 4;
-  
-  /* Slight blur + glow */
-  filter: blur(2px) drop-shadow(0 0 8px rgba(248, 177, 110, 0.411));
-  
-  /* Faint overlay */
-  background: rgba(220, 240, 148, 0.068);
-  mix-blend-mode: screen;
-  border-radius: 0;
-}
-
-/* Scanline flicker animation */
-@keyframes flicker {
-  0%, 100% { opacity: 0.95; }
-  50%      { opacity: 0.75; }
-}
-
-/* Scanline color */
-:root {
-  --scanline-color: 0, 255, 0;
-}
