+++ HttpData/AppData/index.html
@@ -1,0 +1,59 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>TARDQUEST</title>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0, viewport-fit=cover">
+    <link rel="stylesheet" href="style.css">
+</head>
+<body>
+    <div class="custom-titlebar">
+        <img class="custom-titlebar-logo" src="../AppData/titlebar/logo.png">
+        <div class="custom-titlebar-name"><i>TARDQUEST - Extra 'Tarded Edition</i></div>
+        <div class="window-controls">
+            <button class="window-control minimize" onclick="window.electronAPI.minimizeWindow()"></button>
+            <button class="window-control fullscreen" onclick="window.electronAPI.toggleFullscreen()"></button>
+            <button class="window-control maximize" onclick="window.electronAPI.maximizeWindow()"></button>
+            <button class="window-control close" onclick="window.electronAPI.closeWindow()"></button>
+        </div>
+    </div>
+    <div class="toolbar-strip">
+        <div class="border-selector">
+            <select class="border-dropdown" id="borderSelect" onchange="changeBorder()">
+                <option value="" disabled selected>Border Art</option>
+                <option value="">None</option>
+                <option value="characters.png">Characters</option>
+                <option value="erok.png">Erok</option>
+                <option value="tile1.gif">tile1</option>
+            </select>
+        </div>
+        <div class="filter-selector">
+            <select class="filter-dropdown" id="filterSelect" onchange="toggleRetroFilter()">
+                <option value="" disabled selected>Retro Filter</option>
+                <option value="off">Off</option>
+                <option value="on">On</option>
+            </select>
+        </div>
+        <div class="online-selector">
+            <select class="online-dropdown" id="onlineSelect">
+                <option value="" disabled selected>Online</option>
+                <option value="tardboard">TardBoard</option>
+            </select>
+        </div>
+    </div>
+    <div class="main-content">
+        <iframe src="../GameData/game.html" frameborder="0" id="gameFrame"></iframe>
+    </div>
+    <div class="crt-filter" id="crtFilter"></div>
+    <div id="leaderboardWindow" class="draggable-window" style="display:none;">
+        <div class="draggable-titlebar" id="leaderboardTitlebar">
+            <span class="draggable-title">TardBoard</span>
+            <button class="window-control close" id="closeLeaderboardBtn" title="Close"></button>
+        </div>
+        <div class="draggable-content" id="leaderboardContent"></div>
+    </div>
+    <script src="../AppData/scripts/titlebar.js"></script>
+    <script src="../AppData/scripts/retrofilter.js"></script>
+    <script src="../AppData/scripts/tardboard.js"></script>
+</body>
+</html>
