+++ HttpData/AppData/style.css
@@ -1,0 +1,369 @@
+html, body {
+    height: 100%;
+    margin: 0;
+    padding: 0;
+    background-color: #000000;
+}
+
+body {
+    overflow: hidden;
+    margin: 0;
+    padding-top: 0;
+    box-sizing: border-box;
+    height: 100vh;
+    height: 100dvh;
+    position: relative;
+    border: solid 4px;
+    border-image: linear-gradient(135deg, #b1b1b1 0%, #eeeeee 70%, #ffffff 100%) 1;
+    user-select: none;
+    font-family: sans-serif;
+}
+
+iframe {
+    width: 696px;
+    height: 646px;
+    aspect-ratio: 696/646;
+    transition: transform 0.3s ease-in-out;
+    transform-origin: center center;
+    border: none;
+    margin: 0;
+}
+
+iframe#gameFrame {
+    width: 696px;
+    height: 646px;
+    display: block;
+    border: none;
+    background: #000;
+    transform-origin: top center;
+}
+
+.main-content {
+    display: flex;
+    justify-content: center;
+    align-items: flex-start;
+    width: 100%;
+    height: calc(100vh - 43px);
+    margin-top: 43px;
+    overflow: hidden;
+}
+
+.toolbar-strip {
+    position: fixed;
+    top: 20px;
+    left: 0;
+    right: 0;
+    height: 25px;
+    background: linear-gradient(to right, #b1b1b1 0%, #eeeeee 100%);
+    display: flex;
+    align-items: center;
+    box-sizing: border-box;
+    z-index: 999;
+    border: inset 2px;
+    margin-left: 4px;
+    margin-right: 4px;
+}
+
+.custom-titlebar {
+    position: fixed;
+    height: 20px;
+    background-image: url("../AppData/titlebar/bg.png");
+    display: flex;
+    align-items: center;
+    -webkit-app-region: drag;
+    top: 0;
+    left: 0;
+    right: 0;
+    z-index: 1000;
+    padding: 0px;
+    user-select: none;
+    box-shadow: #0000006b 0px 0px 0px 5px;
+}
+
+.custom-titlebar-logo {
+    height: 15px;
+    width: auto;
+    margin-left: 4px;
+    margin-right: 5px;
+    flex-shrink: 0;
+    user-select: none;
+    pointer-events: none;
+    display: block;
+    filter: drop-shadow(1px 1px 0px #0000008f);
+}
+
+.custom-titlebar-name {
+    font-size: 15px;
+    font-weight: bold;
+    color: #000000;
+    text-shadow: #000000 0px 1px 0px;
+    user-select: none;
+    white-space: nowrap;
+    margin-left: 0px;
+    pointer-events: none;
+    flex-shrink: 1;
+    min-width: 0;
+}
+
+.window-controls {
+    margin-left: auto;
+    display: flex;
+    -webkit-app-region: no-drag;
+    flex-shrink: 0;
+}
+
+.window-control {
+    width: 20px;
+    height: 20px;
+    box-shadow: #000000 1px 0px 0px;
+    border: none;
+    margin-right: 5px;
+    cursor: pointer;
+    background-size: 20px 20px;
+    background-repeat: no-repeat;
+    background-position: center;
+    transition: background-color 0.1s ease;
+}
+
+.window-control:hover {
+    background-color: rgba(255, 255, 255, 0.1);
+}
+
+.window-control:active {
+    background-color: rgba(255, 255, 255, 0.3);
+    box-shadow: inset 1px 1px 4px #888, 0 1px 0 #000;
+    transform: translateY(1px);
+}
+
+.close:active {
+    box-shadow: inset 6px 6px 6px rgba(231, 110, 110, 0.671), 0 1px 0 #000;
+    transform: translateY(1px);
+}
+.maximize:active {
+    box-shadow: inset 6px 6px 6px rgba(182, 255, 212, 0.589), 0 1px 0 #000;
+    transform: translateY(1px);
+}
+.fullscreen:active {
+    box-shadow: inset 6px 6px 6px rgba(162, 233, 238, 0.589), 0 1px 0 #000;
+    transform: translateY(1px);
+}
+.minimize:active {
+    box-shadow: inset 6px 6px 6px rgba(245, 255, 154, 0.596), 0 1px 0 #000;
+    transform: translateY(1px);
+}
+
+.minimize:hover,
+.maximize:hover,
+.fullscreen:hover {
+    background-color: rgba(255, 255, 255, 0.15);
+}
+
+.minimize:active,
+.maximize:active,
+.fullscreen:active {
+    background-color: rgba(255, 255, 255, 0.3);
+}
+
+.minimize { 
+    background-image: url("../AppData/titlebar//minimize.png");
+}
+.maximize { 
+    background-image: url("../AppData/titlebar/maximize.png");
+}
+.fullscreen { 
+    background-image: url("../AppData/titlebar/fullscreen.png");
+}
+.close { 
+    background-image: url("../AppData/titlebar/close.png");
+}
+
+.border-dropdown,
+.filter-dropdown {
+    height: 20px;
+    font-size: 13px;
+    text-shadow: #000000 1px 0px 0px;
+    border: outset 2px;
+    color: #000000;
+    background-color: #b1b1b1;
+    cursor: pointer;
+    padding: 0px;
+    margin-left: 5px;
+    outline: none;
+    user-select: none;
+    border-radius: 0px;
+}
+.border-dropdown option,
+.filter-dropdown option,
+.online-dropdown option {
+    border-radius: 0px;
+}
+
+.border-selector,
+.filter-selector,
+.online-selector {
+    -webkit-app-region: no-drag;
+    margin-left: 0;
+    padding-left: 0;
+}
+
+
+/* CRT filter */
+.crt-filter {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100vw;
+    height: 100vh;
+    pointer-events: none;
+    z-index: 500;
+    display: none;
+}
+
+.crt-filter.active {
+    display: block;
+}
+
+.crt-filter::before {
+    content: '';
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background: 
+        linear-gradient(
+            transparent 50%, 
+            rgba(173, 255, 173, 0.09) 50%
+        ),
+        linear-gradient(
+            90deg,
+            transparent 50%, 
+            rgba(255, 0, 0, 0.01) 50%,
+            transparent 52%
+        );
+    background-size: 100% 4px, 3px 100%;
+    animation: flicker 0.15s infinite linear alternate;
+    mix-blend-mode: multiply;
+}
+
+.crt-filter::after {
+    content: '';
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background: rgba(0, 255, 0, 0.05);
+    mix-blend-mode: overlay;
+    animation: glow 2s ease-in-out infinite alternate;
+}
+
+@keyframes flicker {
+    0% {
+        opacity: 1;
+    }
+    2% {
+        opacity: 0.95;
+    }
+    4% {
+        opacity: 0.99;
+    }
+    8% {
+        opacity: 1;
+    }
+    70% {
+        opacity: 0.93;
+    }
+    72% {
+        opacity: 0.97;
+    }
+    77% {
+        opacity: 0.99;
+    }
+    100% {
+        opacity: 1;
+    }
+}
+
+@keyframes glow {
+    0% {
+        opacity: 0.6;
+    }
+    100% {
+        opacity: 0.2;
+    }
+}
+
+/* Draggable TardBoard window */
+.online-dropdown {
+    height: 20px;
+    font-size: 13px;
+    text-shadow: #000000 1px 0px 0px;
+    border: outset 2px;
+    color: #000000;
+    background-color: #b1b1b1;
+    cursor: pointer;
+    padding: 0px;
+    margin-left: 5px;
+    outline: none;
+    user-select: none;
+    font-family: sans-serif;
+}
+
+.draggable-window {
+    position: fixed;
+    top: 80px;
+    left: 80px;
+    width: 225px;
+    height: 500px;
+    background: #000000;
+    border-radius: 0px;
+    z-index: 2000;
+    display: flex;
+    flex-direction: column;
+    user-select: none;
+    overflow: hidden;
+}
+
+.draggable-titlebar {
+    height: 20px;
+    background-image: url("../AppData/titlebar/bg.png");
+    color: #000000;
+    display: flex;
+    justify-content: space-between;
+    padding: 0px;
+    cursor: move;
+    font-size: 16px;
+    font-weight: bold;
+    border-bottom: 4px solid white;
+    text-shadow: #000000 1px 0px 0px;
+}
+
+.draggable-titlebar .draggable-title {
+    margin-left: 4px;
+    font-style: italic;
+    margin-top: 3px;
+}
+
+.draggable-titlebar .window-control.close {
+    margin-left: 0px;
+    align-content: right;
+    background-image: url("../AppData/titlebar/close.png");
+    width: 20px;
+    height: 20px;
+}
+
+.draggable-content {
+    flex: 1;
+    overflow: hidden;
+    background: #000000;
+    padding: 0;
+    border: white 4px solid;
+    border-top: none;
+    margin-top: -4px;
+}
+
+
+
+
+
