+++ HttpData/AppData/titlebar/colored/maximize.png
